textoSql/requirements.txt
# requirements.txt

# Servidor y framework de API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Conexión a PostgreSQL
psycopg2-binary>=2.9.7

# Cliente HTTP asíncrono
httpx>=0.25.0

# Para gestionar variables de entorno (.env)
python-dotenv>=1.0.0